remove old homepage
