+++ app/src/components/landing/Home/index.scss
